make navigation compoents
This will amke navigation a component.
